requirements.txt
fastapi==0.115.12
Pillow==11.2.1
protobuf==6.31.1
python-dotenv==1.1.0
uvicorn==0.34.2
